feat(frontend-sv): add deploy-cloudrun-sv.yml
