+++ frontend-sv/package.json
@@ -1,55 +1,56 @@
 {
-  "name": "frontend-sv",
-  "private": true,
-  "version": "0.0.1",
-  "type": "module",
-  "scripts": {
-    "dev": "vite dev",
-    "build": "vite build",
-    "preview": "vite preview",
-    "prepare": "svelte-kit sync || echo ''",
-    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
-  },
-  "devDependencies": {
-    "@dnd-kit-svelte/core": "^0.0.8",
-    "@dnd-kit-svelte/modifiers": "^0.0.8",
-    "@dnd-kit-svelte/sortable": "^0.0.8",
-    "@dnd-kit-svelte/utilities": "^0.0.8",
-    "@lucide/svelte": "^0.513.0",
-    "@sveltejs/adapter-vercel": "^5.7.2",
-    "@sveltejs/kit": "^2.21.5",
-    "@sveltejs/vite-plugin-svelte": "^5.1.0",
-    "@tabler/icons-svelte": "^3.34.0",
-    "@tailwindcss/forms": "^0.5.10",
-    "@tailwindcss/typography": "^0.5.16",
-    "@tailwindcss/vite": "^4.1.10",
-    "@tanstack/table-core": "^8.21.3",
-    "@types/d3-scale": "^4.0.9",
-    "@types/d3-shape": "^3.1.7",
-    "bits-ui": "^2.7.0",
-    "clsx": "^2.1.1",
-    "d3-scale": "^4.0.2",
-    "d3-shape": "^3.2.0",
-    "formsnap": "^2.0.1",
-    "layerchart": "2.0.0-next.6",
-    "mode-watcher": "^1.0.8",
-    "svelte": "^5.34.3",
-    "svelte-check": "^4.2.1",
-    "svelte-sonner": "^1.0.5",
-    "sveltekit-superforms": "^2.26.1",
-    "tailwind-merge": "^3.3.1",
-    "tailwind-variants": "^1.0.0",
-    "tailwindcss": "^4.1.10",
-    "tw-animate-css": "^1.3.4",
-    "typescript": "^5.8.3",
-    "vaul-svelte": "^1.0.0-next.7",
-    "vite": "^6.3.5"
-  },
-  "dependencies": {
-    "@internationalized/date": "^3.8.2",
-    "date-fns": "^4.1.0",
-    "dotenv": "^16.5.0",
-    "svelte-toolbelt": "^0.9.1"
-  }
+   "name": "frontend-sv",
+   "private": true,
+   "version": "0.0.1",
+   "type": "module",
+   "scripts": {
+      "start": "bun run ./build",
+      "dev": "vite dev",
+      "build": "vite build",
+      "preview": "vite preview",
+      "prepare": "svelte-kit sync || echo ''",
+      "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+      "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
+   },
+   "devDependencies": {
+      "@dnd-kit-svelte/core": "^0.0.8",
+      "@dnd-kit-svelte/modifiers": "^0.0.8",
+      "@dnd-kit-svelte/sortable": "^0.0.8",
+      "@dnd-kit-svelte/utilities": "^0.0.8",
+      "@lucide/svelte": "^0.513.0",
+      "@sveltejs/kit": "^2.21.5",
+      "@sveltejs/vite-plugin-svelte": "^5.1.0",
+      "@tabler/icons-svelte": "^3.34.0",
+      "@tailwindcss/forms": "^0.5.10",
+      "@tailwindcss/typography": "^0.5.16",
+      "@tailwindcss/vite": "^4.1.10",
+      "@tanstack/table-core": "^8.21.3",
+      "@types/d3-scale": "^4.0.9",
+      "@types/d3-shape": "^3.1.7",
+      "bits-ui": "^2.7.0",
+      "clsx": "^2.1.1",
+      "d3-scale": "^4.0.2",
+      "d3-shape": "^3.2.0",
+      "formsnap": "^2.0.1",
+      "layerchart": "2.0.0-next.6",
+      "mode-watcher": "^1.0.8",
+      "svelte": "^5.34.3",
+      "svelte-adapter-bun": "^0.5.2",
+      "svelte-check": "^4.2.1",
+      "svelte-sonner": "^1.0.5",
+      "sveltekit-superforms": "^2.26.1",
+      "tailwind-merge": "^3.3.1",
+      "tailwind-variants": "^1.0.0",
+      "tailwindcss": "^4.1.10",
+      "tw-animate-css": "^1.3.4",
+      "typescript": "^5.8.3",
+      "vaul-svelte": "^1.0.0-next.7",
+      "vite": "^6.3.5"
+   },
+   "dependencies": {
+      "@internationalized/date": "^3.8.2",
+      "date-fns": "^4.1.0",
+      "dotenv": "^16.5.0",
+      "svelte-toolbelt": "^0.9.1"
+   }
 }
