+++ frontend-sv/src/routes/+layout.server.ts
@@ -1,25 +1,25 @@
-import type { LayoutServerLoad } from './$types';
+// import type { LayoutServerLoad } from './$types';
 
-export const load: LayoutServerLoad = async ({ cookies }) => {
-  // this is stupid use locals
-  // Get user data from cookie
-  const userCookie = cookies.get('user');
-  const accessToken = cookies.get('access_token');
+// export const load: LayoutServerLoad = async ({ cookies }) => {
+//   // this is stupid use locals
+//   // Get user data from cookie
+//   const userCookie = cookies.get('user');
+//   const accessToken = cookies.get('access_token');
 
-  let user = null;
+//   let user = null;
 
-  if (userCookie) {
-    try {
-      user = JSON.parse(userCookie);
-    } catch (error) {
-      console.error('Error parsing user cookie:', error);
-      // Clear invalid cookie
-      cookies.delete('user', { path: '/' });
-    }
-  }
+//   if (userCookie) {
+//     try {
+//       user = JSON.parse(userCookie);
+//     } catch (error) {
+//       console.error('Error parsing user cookie:', error);
+//       // Clear invalid cookie
+//       cookies.delete('user', { path: '/' });
+//     }
+//   }
 
-  return {
-    user,
-    isAuthenticated: !!accessToken && !!user
-  };
-}
+//   return {
+//     user,
+//     isAuthenticated: !!accessToken && !!user
+//   };
+// }
