+++ frontend-sv/svelte.config.js
@@ -1,5 +1,5 @@
-import adapter from '@sveltejs/adapter-vercel';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
+import adapter from 'svelte-adapter-bun';
 
 const config = {
   preprocess: vitePreprocess(),
